move: participant component
