+++ src/main/vue/src/assets/css/member/profile-modal.css
@@ -1,0 +1,322 @@
+#sheet {
+  --overlay: #888;
+  --divider: #333;
+  position: fixed;
+  top: 0;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  z-index: 300;
+  visibility: visible;
+  transition: opacity 0.5s, visibility 0.5s;
+}
+
+#sheet[aria-hidden="true"] {
+  opacity: 0;
+  visibility: hidden;
+  pointer-events: none;
+}
+
+#sheet .contents:not(.not-selectable) {
+  transition: var(--default-transitions), height 0.5s;
+}
+
+
+#sheet[aria-hidden="true"] .contents {
+  transform: translateY(100%);
+}
+
+#sheet .overlay {
+  position: absolute;
+  top: 0;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  z-index: -1;
+  background: var(--overlay);
+  opacity: 0.5;
+}
+
+#sheet .contents {
+  border-radius: 1rem 1rem 0 0;
+  position: relative;
+  overflow-y: hidden;
+
+  /* 임우빈의 수정사항  css */
+  /* top:570px;
+  max-width: 400px;
+  margin : auto;  */
+  /* 임우빈 수정사항 끝 */
+
+  --default-transitions: transform 0.5s, border-radius 0.5s;
+  background-color: #fff;
+  transition: var(--default-transitions);
+  transform: translateY(0);
+  max-height: 42vh;
+  height: 42vh;
+  min-width: 360px;
+  box-sizing: border-box;
+  padding-top:40px;
+  padding-left: 18px;
+  padding-right :18px;
+  padding-bottom: 5px;
+}
+
+#sheet .draggable-area {
+  position: absolute;
+  top: 0;
+  left: 0;
+  right: 0;
+  width: 3rem;
+  margin: auto;
+  padding-top: 10px;
+  padding-bottom: 18px;
+  cursor: grab;
+}
+
+#sheet .draggable-thumb {
+  width: inherit;
+  height: 0.25rem;
+  background: var(--divider);
+  border-radius: 0.125rem;
+}
+
+#sheet .close-sheet {
+  position: absolute;
+  right: 0;
+  top: 0;
+  border: none;
+  margin-right: 15px;
+  margin-top: 10px;
+}
+
+#sheet .body {
+  height: 100%;
+  overflow-y: hidden;
+  gap: 1rem;
+}
+
+.resigned-member{
+  height: 20px;
+  color:red;
+  padding-top:50px;
+}
+
+.image { 
+  width: 100px;
+  height: 100px;
+  border-radius: 99px;
+  position: absolute;
+ left: 2px;
+ top: 2px;
+}
+
+.image-wrap {
+  display: flex;
+  justify-content: center;
+
+}
+.image-bg{
+  width: 104px;
+  height: 104px;
+  background-color: var(--tiffany-blue);
+  position: relative;
+  border-radius: 99px;
+}
+.user__info{
+  font-size: 16px;
+  display: flex;
+  justify-content: center;
+  margin-top: 23px;
+  line-height: 20px;
+}
+
+.user__info-fixed{
+  display: flex;
+  flex-direction: column;
+  color: var(--main-color);
+  font-weight: 600;
+  margin-left: -30px;
+  margin-right: 10px;
+}
+.user__info-flexible{
+  color: var(--dark-grey2);
+  display: flex;
+  flex-direction: column;
+  padding-left: 10px;
+  padding-right:10px;
+}
+
+.report{
+  padding-left: 10px;
+  background-color: var(--main-color)
+}
+.kick{
+  padding-left: 10px;
+  background-color: var(--tiffany-blue)
+}
+
+span.icon{
+margin-right: 10px;
+}
+.ban__wrap{
+  display: flex;
+  justify-content: center;
+  margin-top: 26px;
+  gap:10px;
+}
+
+
+
+@media (min-width: 576px){
+  .btn__wrap.btn {
+    padding-left: 10px;
+    padding-right: 10px;
+}
+}
+
+.row {
+  display: flex;
+  flex-direction: row;
+}
+.row.reversed-order {
+  flex-direction: row-reverse;
+}
+
+.column {
+  display: flex;
+  flex-direction: column;
+}
+.column.reversed-order {
+  flex-direction: column-reverse;
+}
+
+.row.items-start,
+.column.items-start {
+  align-items: flex-start;
+}
+.row.justify-start,
+.column.justify-start {
+  justify-content: flex-start;
+}
+.row.content-start,
+.column.content-start {
+  align-content: flex-start;
+}
+.row.items-center,
+.column.items-center {
+  align-items: center;
+}
+.row.justify-center,
+.column.justify-center {
+  justify-content: center;
+}
+.row.content-center,
+.column.content-center {
+  align-content: center;
+}
+.row.items-end,
+.column.items-end {
+  align-items: flex-end;
+}
+.row.justify-end,
+.column.justify-end {
+  justify-content: flex-end;
+}
+.row.content-end,
+.column.content-end {
+  align-content: flex-end;
+}
+.row.items-stretch,
+.column.items-stretch {
+  align-items: stretch;
+}
+.row.justify-stretch,
+.column.justify-stretch {
+  justify-content: stretch;
+}
+.row.content-stretch,
+.column.content-stretch {
+  align-content: stretch;
+}
+.row.items-baseline,
+.column.items-baseline {
+  align-items: baseline;
+}
+.row.justify-baseline,
+.column.justify-baseline {
+  justify-content: baseline;
+}
+.row.content-baseline,
+.column.content-baseline {
+  align-content: baseline;
+}
+.row .wrap,
+.column .wrap {
+  flex-wrap: wrap;
+}
+.row .reversed-wrap,
+.column .reversed-wrap {
+  flex-wrap: wrap-reverse;
+}
+.row .no-wrap,
+.column .no-wrap {
+  flex-wrap: nowrap;
+}
+
+.fill {
+  flex-grow: 1;
+  flex-shrink: 0;
+}
+
+.font-system {
+  font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, Ubuntu, Helvetica, Arial, sans-serif, "Apple Color Emoji", "Segoe UI Emoji", "Segoe UI Symbol";
+}
+
+.font-thin {
+  font-weight: 100;
+}
+
+.font-extra-light {
+  font-weight: 200;
+}
+
+.font-light {
+  font-weight: 300;
+}
+
+.font-regular {
+  font-weight: 400;
+}
+
+.font-medium {
+  font-weight: 500;
+}
+
+.font-semibold {
+  font-weight: 600;
+}
+
+.font-bold {
+  font-weight: 700;
+}
+
+.font-extra-bold {
+  font-weight: 800;
+}
+
+.font-black {
+  font-weight: 900;
+}
+
+.cursor-pointer {
+  cursor: pointer;
+}
+
+.not-selectable {
+  user-select: none;
+}
+
+.selectable {
+  user-select: auto;
+}
