+++ src/main/vue/src/components/meeting/participant/MemberProfile.vue
@@ -1,6 +1,7 @@
 <script setup>
 import { reactive, ref } from "vue";
-import ReportUser from "../member/ReportUser.vue";
+import ReportUser from "@/components/member/ReportUser.vue";
+import DefaultModal from "@/components/modal/Default.vue";
 import { useMeetingDetailStore } from "@/stores/meetingDetailStore";
 import { useMemberStore } from "@/stores/memberStore";
 import api from "@/api";
@@ -100,10 +101,7 @@ async function onClickKick() {
         <div class="draggable-area">
           <div class="draggable-thumb"></div>
         </div>
-        <span
-          @click.prevent="$emit('closeModal')"
-          class="close-sheet icon icon-x"
-        ></span>
+        <span @click.prevent="$emit('closeModal')" class="close-sheet icon icon-x"></span>
       </header>
 
       <div class="body column">
@@ -111,21 +109,13 @@ async function onClickKick() {
         <div v-if="state.member.resigned" class="profile-container">
           <div class="image-wrap">
             <div class="image-bg">
-              <img
-                src="/images/no-rusult-ghost.svg"
-                class="image"
-                alt="resigned"
-              />
+              <img src="/images/no-rusult-ghost.svg" class="image" alt="resigned" />
             </div>
           </div>
           <div class="resigned">
             <span>탈퇴한 회원이에요</span>
             <div class="ban__wrap">
-              <span
-                v-show="kickOn"
-                @click="onClickKick"
-                class="btn btn-semiround kick"
-              >
+              <span v-show="kickOn" @click="onClickKick" class="btn btn-semiround kick">
                 <span class="icon icon-door"></span>
                 내보내기
               </span>
@@ -171,11 +161,7 @@ async function onClickKick() {
               <span class="icon icon-siren-white"></span>
               신고하기
             </span>
-            <span
-              v-show="kickOn"
-              @click="onClickKick"
-              class="btn btn-semiround kick"
-            >
+            <span v-show="kickOn" @click="onClickKick" class="btn btn-semiround kick">
               <span class="icon icon-door"></span>
               내보내기
             </span>
@@ -185,10 +171,11 @@ async function onClickKick() {
     </div>
   </div>
   <ReportUser @closeModal="closeModal" v-if="reportModalOn" />
+  <!-- <default-modal></default-modal> -->
 </template>
 
 <style scoped>
-@import url(../../../../resources/static/css/member/member/profile-modal.css);
+@import url(@/assets/css/member/profile-modal.css);
 .resigned {
   margin-top: 10px;
   display: flex;
