+++ src/main/vue/src/components/meeting/participant/Participant.vue
